Message utils (#10)
* :sparkles: - database setup done

* :sparkles: - Message utils added

* :sparkles: - Message utils updated

* :sparkles: - Message utils updated

---------
